pyproject.toml
[project]
name = "dbt-graphql"
version = "0.1.0"
description = "Convert dbt artifacts to GraphQL schema and serve a SQL-backed GraphQL API"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = "MIT"
dependencies = [
    "dbt-artifacts-parser>=0.13.0",
    "dbt-colibri>=0.3.0,<0.4",
    "graphql-core>=3.2.8",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "opentelemetry-exporter-otlp>=1.41.0",
    "opentelemetry-sdk>=1.41.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0.2",
    "simpleeval>=1.0.7",
    "sqlalchemy[asyncio]>=2.0.49",
    "cashews>=7.0",
]

[project.optional-dependencies]
postgres = ["asyncpg>=0.31"]
mysql = ["aiomysql>=0.3.2"]
api = [
    "ariadne>=1.0.1",
    "granian>=2.7.3",
    "pyjwt>=2.12.1",
    "opentelemetry-instrumentation-sqlalchemy>=0.62b0",
    "opentelemetry-instrumentation-starlette>=0.62b0",
]
mcp = ["fastmcp>=3.2.4"]

[project.scripts]
dbt-graphql = "dbt_graphql.cli:main"

[dependency-groups]
test = ["pytest", "pytest-cov", "pytest-asyncio", "pytest-docker", "httpx"]
lint = ["ruff", "ty"]

[build-system]
requires = ["uv_build>=0.11.6,<0.12"]
build-backend = "uv_build"


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
omit = ["tests/*", "**/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
